# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Midas DCF Valuation API
  version: 0.1.0
  description: |
    REST API exposing intrinsic value metrics per ticker.
    - Net Tangible Asset Value per share
    - DCF Fair Value per share (5-year projection, WACC-discounted)
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/fair-value/{ticker}:
    get:
      summary: Get fair value for a ticker
      tags: [valuation]
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 5
        - name: override_beta
          in: query
          required: false
          schema:
            type: number
            format: float
        - name: override_rf
          in: query
          required: false
          schema:
            type: number
            format: float
      responses:
        '200':
          description: Fair value metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FairValueResponse'
        '400':
          description: Invalid ticker or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ticker not present in SEC's ticker→CIK index (genuinely unknown symbol)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ticker resolves but cannot be valued — e.g. SEC has no usable US-GAAP XBRL facts (foreign private issuers, some pre-revenue issuers) or no model is applicable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/fair-value/bulk:
    post:
      summary: Get fair value for multiple tickers
      tags: [valuation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkFairValueRequest'
      responses:
        '200':
          description: Bulk fair value results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkFairValueResponse'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Industry:
      type: object
      description: |
        Dual industry classification — both the SIC-derived label (used
        internally to select the valuation model) and the balance-sheet
        heuristic label (used by the datacleaner). The `match` flag is true
        when the two classifiers agree per the canonical SIC -> GICS mapping,
        and false otherwise. A false match signals potential classifier drift
        worth investigating; see
        docs/superpowers/specs/2026-04-23-industry-in-response-design.md.
      properties:
        sic_code:
          type: string
          description: Raw SIC code from SEC (omitted if unavailable)
          example: "3674"
        sic:
          type: string
          description: SIC-derived industry label (TECH, MFG, RETAIL, UTIL, FINL, HEALTH, NA)
          example: "MFG"
        heuristic_code:
          type: string
          description: GICS sector code from the balance-sheet heuristic classifier
          example: "45"
        heuristic_name:
          type: string
          description: Human-readable GICS sector name
          example: "Information Technology"
        match:
          type: boolean
          description: True when SIC and heuristic agree per the canonical mapping
          example: true
    FairValueResponse:
      type: object
      properties:
        ticker:
          type: string
        wacc:
          type: number
          format: float
        growth_rate:
          type: number
          format: float
        tangible_value_per_share:
          type: number
          format: float
        dcf_value_per_share:
          type: number
          format: float
        as_of:
          type: string
          format: date-time
        data_quality_score:
          type: number
          format: float
        data_quality_grade:
          type: string
        industry:
          $ref: '#/components/schemas/Industry'
    BulkFairValueRequest:
      type: object
      required: [tickers]
      properties:
        tickers:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: string
        override_beta:
          type: number
          format: float
        override_rf:
          type: number
          format: float
    BulkSummary:
      type: object
      properties:
        total_requested:
          type: integer
        successful:
          type: integer
        failed:
          type: integer
    BulkFairValueResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/FairValueResponse'
        summary:
          $ref: '#/components/schemas/BulkSummary'
    ErrorResponse:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        context:
          type: object
          additionalProperties: true